+++ src/shared.interfaces.ts
@@ -2,14 +2,17 @@ import { StaticConfig } from './JsonSettings.interface';
 import { ApplicationConfiguration } from './YamlSettings.interface';
 
 export interface BackgroundApi {
+  autostartTaskExist: () => Promise<boolean>;
+  // actions
   enableAutostart: () => void;
   disableAutostart: () => void;
-  autostartTaskExist: () => Promise<boolean>;
+  forceRestart: () => void;
+  quit: () => void;
+  // settings
   getUserSettings: (route?: string) => Promise<UserSettings>;
   saveUserSettings: (settings: UserSettings) => Promise<void>;
   loadAppsTemplate: () => Promise<ApplicationConfiguration[]>;
   exportAppsTemplate: (apps: ApplicationConfiguration[]) => Promise<void>;
-  quit: () => void;
   // installers
   runAhkSetup: () => void;
 }