+++ static/pwsh/force_stop.ps1
@@ -1,9 +1,14 @@
-param (
-  [string]$ExeRoute
-)
+# Stop Task if exist
+$task = Get-ScheduledTask -TaskName KomorebiUI -ErrorAction SilentlyContinue
+if ($null -ne $task -and $task.State -eq "Running") { 
+  Stop-ScheduledTask -TaskName KomorebiUI
+}
 
-$runningProcesses = Get-Process | Where-Object { $_.Path -eq $ExeRoute }
+# Stop AHK script
+wmic process where "commandline like '%komorebic.ahk%'" call terminate
+
+# Check for others komorebi process caused by others sources.
+$runningProcesses = Get-Process | Where-Object { $_.ProcessName -eq "komorebi" }
 if ($runningProcesses.Count -gt 0) {
-  Stop-Process -Name komorebi
-  Stop-Process -Name "komorebi-ui"
+  Stop-Process -Id $runningProcesses.Id
 }