+++ static/pwsh/manual_run.ps1
@@ -2,7 +2,13 @@ param (
   [string]$ExeRoute
 )
 
-$runningProcesses = Get-Process | Where-Object { $_.Path -eq $ExeRoute }
+$runningProcesses = Get-Process | Where-Object { $_.ProcessName -eq "komorebi" }
 if ($runningProcesses.Count -eq 0) {
-  Start-Process powershell -WindowStyle Hidden -ArgumentList "-NoProfile -Command `"& '$ExeRoute' -c '$Env:USERPROFILE\.config\komorebi-ui\settings.json'`""
+  $task = Get-ScheduledTask -TaskName KomorebiUI -ErrorAction SilentlyContinue
+  if ($null -ne $task) { 
+    Start-ScheduledTask -TaskName KomorebiUI
+  } else {
+    Start-Process powershell -WindowStyle Hidden -ArgumentList "-NoProfile -Command `"& '$ExeRoute' -c '$Env:USERPROFILE\.config\komorebi-ui\settings.json'`""
+    & (Join-Path $Env:USERPROFILE "\.config\komorebi-ui\komorebic.ahk")
+  }
 }